Allow opening sub-page.
When integrating other websites in iframes and indexing those
contents in the search, we want to be able to link from the search
to a specific iframed sub-page.
In order to make this possible we need to be able to pass the
requested sub-page as GET request param.
For security reason, the origin of both URLs must be the same,
otherwise the configured startpage is loaded.