Load iframeResize later.
If the iframeResize function is called too early the postMessage will return an error that the origin and target do not match. If the method is called later (in the onload of the iframe) the error is not raised anymore. Also see https://github.com/davidjbradshaw/iframe-resizer/issues/443#issuecomment-331721886 as a reference.